*More code cleanup.  Although I don't usually bother with the third party source files, I removed unused variables from some of the AGUP ones because they now get warnings in the latest GCC, which would have annoyed me to no end. *Updated GH import so that pitch values of 2 are treated as pitch shifts. *Updated INI export so that the pro drums not present tag is written if no pro drum difficulty is defined and there are no cymbals, and a vocal harmonies not present tag if no vocal harmony difficulty is defined and there are no harmony tracks stored into the project.

+++ trunk/src/editor.h
@@ -44,7 +44,7 @@ typedef struct
 extern EOF_SNAP_DATA eof_snap;
 extern EOF_SNAP_DATA eof_tail_snap;
 
-void eof_select_beat(int beat);
+void eof_select_beat(unsigned long beat);
 void eof_snap_logic(EOF_SNAP_DATA * sp, unsigned long p);
 void eof_snap_length_logic(EOF_SNAP_DATA * sp);	//Calculates the grid snap interval length for sp->beat
 void eof_read_editor_keys(void);
